requirements.txt
django==1.11.6
mongoengine==0.13.0
pymongo==3.4.0
#django-configurations
django-redis==4.8.0
django-cors-headers==2.1.0
django-el-pagination==3.1.0
ujson==1.35 

uWSGI==2.0.15
requests==2.18.4
python-dateutil==2.6.1
pyotp==2.2.1
razorpay==1.2.0
cryptography==2.6.1
sentry-sdk==1.0.0